Modified vaccination feature
