+++ App.js
@@ -12,7 +12,6 @@ import VerificationCodeComponents from './login_feature/VerificationCodeScreen.j
 import StartScreenComponents from './login_feature/StartScreen.js';
 import SignUpComponents from './login_feature/SignUp.js';
 import VerificationEmailComponents from './login_feature/VerificationEmail.js';
-import VaccineApp from './vaccination_feature/VaccinationScreen.js';
 import Vaccines from './vaccination_feature/VaccineListScreen.js';
 
 const Stack = createStackNavigator();
@@ -41,23 +40,7 @@ const App = () => {
                               )
                     })} /> 
 
-          <Stack.Screen 
-                name="Vaccine" 
-                component={VaccineApp}
-                  options={({ navigation }) => ({
-                  title: 'Vaccination',
-                  headerTitleAlign: 'center',
-                  headerTintColor: '#FFFFFF', // Change the text color of the header title
-                  headerStyle: { backgroundColor: '#649F95',height: 100,},
-                  headerTitleStyle: { fontSize: 24,  }, // Set the font size and weight of the header title // Set the background color of the header // Align the title to the center
-                      headerLeft: () => ( // Add custom headerLeft component
-                                  <TouchableOpacity onPress={() => navigation.goBack()}>
-                                    <View style={styles.backButton}>
-                                    <FontAwesomeIcon icon={faChevronLeft} size={24} color="#FFFFFF" />
-                                    </View>                         
-                                  </TouchableOpacity>
-                              )
-                    })} /> 
+          
 
           <Stack.Screen 
                   name="StartScreen"
