+++ package.json
@@ -14,6 +14,7 @@
     "@fortawesome/react-native-fontawesome": "^0.3.0",
     "@react-native-community/blur": "^4.4.0",
     "@react-native-community/checkbox": "^0.5.17",
+    "@react-native-community/datetimepicker": "^7.6.2",
     "@react-navigation/native": "^6.1.10",
     "@react-navigation/stack": "^6.3.21",
     "expo": "~50.0.5",
